requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
aliyun-python-sdk-core>=2.16.0
oss2>=2.19.0
python-dotenv>=1.0.0
httpx>=0.27.0
